week1/requirements.txt
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-storage
gunicorn
mesop
vertexai
